+++ Default.sublime-keymap
@@ -1,43 +1,4 @@
 [
-    // TODO: Only for testing.
-    {
-        "keys": [
-            "left"
-        ],
-        "command": "lsp_move_focus",
-        "args": {
-            "direction": "left"
-        },
-    },
-    {
-        "keys": [
-            "right"
-        ],
-        "command": "lsp_move_focus",
-        "args": {
-            "direction": "right"
-        },
-    },
-    {
-        "keys": [
-            "up"
-        ],
-        "command": "lsp_move_focus",
-        "args": {
-            "direction": "up"
-        },
-    },
-    {
-        "keys": [
-            "down"
-        ],
-        "command": "lsp_move_focus",
-        "args": {
-            "direction": "down"
-        },
-    },
-
-
     // To enable or change an existing key binding, copy it to the right view and uncomment, if needed.
     // To change the associated key press, edit the "keys" value.
 
@@ -282,5 +243,36 @@
         "command": "lsp_save",
         "args": {"async": true},
         "context": [{"key": "lsp.session_with_capability", "operand": "textDocumentSync.willSave | textDocumentSync.willSaveWaitUntil | codeActionProvider.codeActionKinds | documentFormattingProvider | documentRangeFormattingProvider"}]
+    },
+    // Navigate in Tree Views
+    {
+        "keys": ["left"],
+        "command": "lsp_handle_tree_view_action",
+        "args": {"action": "move_left"},
+        "context": [{"key": "lsp.tree_view_sheet_has_focus"}]
+    },
+    {
+        "keys": ["right"],
+        "command": "lsp_handle_tree_view_action",
+        "args": {"action": "move_right"},
+        "context": [{"key": "lsp.tree_view_sheet_has_focus"}]
+    },
+    {
+        "keys": ["up"],
+        "command": "lsp_handle_tree_view_action",
+        "args": {"action": "move_up"},
+        "context": [{"key": "lsp.tree_view_sheet_has_focus"}]
+    },
+    {
+        "keys": ["down"],
+        "command": "lsp_handle_tree_view_action",
+        "args": {"action": "move_down"},
+        "context": [{"key": "lsp.tree_view_sheet_has_focus"}]
+    },
+    {
+        "keys": ["escape"],
+        "command": "lsp_handle_tree_view_action",
+        "args": {"action": "close"},
+        "context": [{"key": "lsp.tree_view_sheet_has_focus"}]
     },
 ]
